+++ lib/src/annotation/polygon_annotation_manager.dart
@@ -85,6 +85,18 @@ class PolygonAnnotationManager extends BaseAnnotationManager {
   /// Delete all the annotation added by this manager.
   Future<void> deleteAll() => _annotationMessenger.deleteAll(id);
 
+  /// Determines whether bridge guard rails are added for elevated roads. Default value: "true".
+  @experimental
+  Future<void> setFillConstructBridgeGuardRail(
+          bool fillConstructBridgeGuardRail) =>
+      _annotationMessenger.setFillConstructBridgeGuardRail(
+          id, fillConstructBridgeGuardRail);
+
+  /// Determines whether bridge guard rails are added for elevated roads. Default value: "true".
+  @experimental
+  Future<bool?> getFillConstructBridgeGuardRail() =>
+      _annotationMessenger.getFillConstructBridgeGuardRail(id);
+
   /// Selects the base of fill-elevation. Some modes might require precomputed elevation data in the tileset. Default value: "none".
   @experimental
   Future<void> setFillElevationReference(
@@ -110,6 +122,17 @@ class PolygonAnnotationManager extends BaseAnnotationManager {
 
   /// Whether or not the fill should be antialiased. Default value: true.
   Future<bool?> getFillAntialias() => _annotationMessenger.getFillAntialias(id);
+
+  /// The color of bridge guard rail. Default value: "rgba(241, 236, 225, 255)".
+  @experimental
+  Future<void> setFillBridgeGuardRailColor(int fillBridgeGuardRailColor) =>
+      _annotationMessenger.setFillBridgeGuardRailColor(
+          id, fillBridgeGuardRailColor);
+
+  /// The color of bridge guard rail. Default value: "rgba(241, 236, 225, 255)".
+  @experimental
+  Future<int?> getFillBridgeGuardRailColor() =>
+      _annotationMessenger.getFillBridgeGuardRailColor(id);
 
   /// The color of the filled part of this layer. This color can be specified as `rgba` with an alpha component and the color's opacity will not affect the opacity of the 1px stroke, if it is used. Default value: "#000000".
   Future<void> setFillColor(int fillColor) =>
@@ -164,6 +187,17 @@ class PolygonAnnotationManager extends BaseAnnotationManager {
   /// Controls the frame of reference for `fill-translate`. Default value: "map".
   Future<FillTranslateAnchor?> getFillTranslateAnchor() =>
       _annotationMessenger.getFillTranslateAnchor(id);
+
+  /// The color of tunnel structures (tunnel entrance and tunnel walls). Default value: "rgba(241, 236, 225, 255)".
+  @experimental
+  Future<void> setFillTunnelStructureColor(int fillTunnelStructureColor) =>
+      _annotationMessenger.setFillTunnelStructureColor(
+          id, fillTunnelStructureColor);
+
+  /// The color of tunnel structures (tunnel entrance and tunnel walls). Default value: "rgba(241, 236, 225, 255)".
+  @experimental
+  Future<int?> getFillTunnelStructureColor() =>
+      _annotationMessenger.getFillTunnelStructureColor(id);
 
   /// Specifies an uniform elevation in meters. Note: If the value is zero, the layer will be rendered on the ground. Non-zero values will elevate the layer from the sea level, which can cause it to be rendered below the terrain. Default value: 0. Minimum value: 0.
   @experimental