+++ lib/src/style/layer/circle_layer.dart
@@ -13,6 +13,8 @@ class CircleLayer extends Layer {
     String? slot,
     required String this.sourceId,
     String? this.sourceLayer,
+    CircleElevationReference? this.circleElevationReference,
+    List<Object>? this.circleElevationReferenceExpression,
     double? this.circleSortKey,
     List<Object>? this.circleSortKeyExpression,
     double? this.circleBlur,
@@ -56,6 +58,16 @@ class CircleLayer extends Layer {
 
   /// A source layer is an individual layer of data within a vector source. A vector source can have multiple source layers.
   String? sourceLayer;
+
+  /// Selects the base of circle-elevation. Some modes might require precomputed elevation data in the tileset.
+  /// Default value: "none".
+  @experimental
+  CircleElevationReference? circleElevationReference;
+
+  /// Selects the base of circle-elevation. Some modes might require precomputed elevation data in the tileset.
+  /// Default value: "none".
+  @experimental
+  List<Object>? circleElevationReferenceExpression;
 
   /// Sorts features in ascending order based on this value. Features with a higher sort key will appear above features with a lower sort key.
   double? circleSortKey;
@@ -170,6 +182,14 @@ class CircleLayer extends Layer {
           visibility!.name.toLowerCase().replaceAll("_", "-");
     }
 
+    if (circleElevationReferenceExpression != null) {
+      layout["circle-elevation-reference"] = circleElevationReferenceExpression;
+    }
+
+    if (circleElevationReference != null) {
+      layout["circle-elevation-reference"] =
+          circleElevationReference?.name.toLowerCase().replaceAll("_", "-");
+    }
     if (circleSortKeyExpression != null) {
       layout["circle-sort-key"] = circleSortKeyExpression;
     }
@@ -314,6 +334,15 @@ class CircleLayer extends Layer {
               .contains(map["layout"]["visibility"])),
       visibilityExpression: _optionalCastList(map["layout"]["visibility"]),
       filter: _optionalCastList(map["filter"]),
+      circleElevationReference:
+          map["layout"]["circle-elevation-reference"] == null
+              ? null
+              : CircleElevationReference.values.firstWhere((e) => e.name
+                  .toLowerCase()
+                  .replaceAll("_", "-")
+                  .contains(map["layout"]["circle-elevation-reference"])),
+      circleElevationReferenceExpression:
+          _optionalCastList(map["layout"]["circle-elevation-reference"]),
       circleSortKey: _optionalCast(map["layout"]["circle-sort-key"]),
       circleSortKeyExpression:
           _optionalCastList(map["layout"]["circle-sort-key"]),